requirements.txt
beautifulsoup4==4.11.2
BitstampClient==2.2.10
cbpro==1.1.4
ccxt==2.8.94
pandas==1.5.3
plotly==5.13.1
pycoingecko==3.1.0
pytest==7.2.1
python-decouple==3.8
requests==2.28.2
websockets==10.4
wget==3.2
